Move `sub_500907` to C++

+++ src/server.h
@@ -82,7 +82,7 @@ public:
 
     // Sub-functions called by sub_4FC644
     Human* sub_4EF4E7(void* block, int arg14, Player* player); // Load character.
-    Human* sub_500907(Player* player, uint8_t body, uint8_t reaction, uint8_t mind, uint8_t spirit, uint8_t main_skill, uint32_t character_class);
+    Human* sub_500907(Player* player, uint8_t body, uint8_t reaction, uint8_t mind, uint8_t spirit, uint8_t main_skill, uint8_t character_class);
     void sub_4EE028(Unit* unit);
     void sub_5013D4(Player* player);
     void sub_4F4570();
